created dateConversion and moved getMonthLength, getMonthString, and checkforLeapYear into that module. Likewise with dateConversion_test. All tests still passing
