+++ lib/render.js
@@ -1,26 +1,13 @@
 'use strict';
 
-module.exports.getMonthString = (monthNum) => {
-  const month = {
-    1: 'January',
-    2: 'February',
-    3: 'March',
-    4: 'April',
-    5: 'May',
-    6: 'June',
-    7: 'July',
-    8: 'August',
-    9: 'September',
-    10: 'October',
-    11: 'November',
-    12: 'December'
-  }
-
-  return month[monthNum]
-}
+const { 
+  getMonthString,
+  getMonthLength,
+  checkForLeapYear
+} = require('./dateConversion')
 
 module.exports.renderCal = (dayOfWeek, dateObj) => {
-  let monthString = module.exports.getMonthString(dateObj.month) + ` ${dateObj.year}`
+  let monthString = getMonthString(dateObj.month) + ` ${dateObj.year}`
   let mainHeader = module.exports.centerString(monthString, 20, false)
   let fixedHeader = module.exports.createFixedHeader()
   let weeks = module.exports.createWeeks(dayOfWeek, dateObj.month, dateObj.year)
@@ -35,7 +22,7 @@ module.exports.renderCal = (dayOfWeek, dateObj) => {
 }
 
 module.exports.renderYearCal = (dayOfWeek, dateObj) => {
-  let monthString = module.exports.getMonthString(dateObj.month)
+  let monthString = getMonthString(dateObj.month)
   let mainHeader = module.exports.centerString(monthString, 20, false)
   let fixedHeader = module.exports.createFixedHeader()
   let weeks = module.exports.createWeeks(dayOfWeek, dateObj.month, dateObj.year)
@@ -80,7 +67,7 @@ module.exports.createWeeks = (dayOfWeek, month, year) => {
   let grids = 42
   const PADDING = ' '
   const EMPTY_GRID = '  '
-  const monthLength = module.exports.getMonthLength(month, year)
+  const monthLength = getMonthLength(month, year)
 
   let weekString = ''
 
@@ -151,32 +138,4 @@ module.exports.newLine = (array) => {
   string += '\n'
   newLine.push(string)
   return newLine
-}
-
-module.exports.checkForLeapYear = (year) => {
-  return ((year % 4 === 0 && year % 100 !== 0) || (year % 400 === 0))
-}
-
-module.exports.getMonthLength = (month, year) => {
-  let isLeapYear = module.exports.checkForLeapYear(year)
-  let monthLengths = [
-    31,
-    28,
-    31,
-    30,
-    31,
-    30,
-    31,
-    31,
-    30,
-    31,
-    30,
-    31
-  ]
-  let length = monthLengths[month - 1]
-  if (month == 2 && isLeapYear) {
-    return 29
-  } else {
-    return length
-  }
 }
